Update RevampedCrossWord
The program loops until it finds a full combination. Added checkFuture which checks to see if places a word in a row or column, letters that will add to the column or row it is aligned with can create a future word. Currently only works for nxn crosswords. Lots of not real words, I am going to add a definition requirement.